Ecommerce_Django_App/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-crispy-forms==1.14.0
django-cropperjs==0.0.5
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
twilio==7.9.3
urllib3==1.26.9
whitenoise==6.2.0
